Refactor character stat retrieval and damage calculations into new modules
